fix(decompiler): perform PE bounds checks in 64-bit so they cannot be bypassed by overflow
The e_lfanew range guard in WixBurnAccess.Open computed `eLfanew + 4 >
stream.Length` in unchecked Int32 arithmetic. e_lfanew is read straight from
the (attacker-controlled) PE header, so a value near Int32.MaxValue makes the
addition wrap to a negative number, silently defeating the explicit guard.
The bogus offset then fails downstream when BinaryReader hits EOF, so the
outer catch-all still rejected the file (WBD002) - but only by accident, and
a future refactor that narrows the catch-all would turn this into a real
bug in a parser that reads foreign executables during forge migrate.

Widen the comparison to long (stream.Length is already long) so the guard
rejects on its own, without depending on the downstream exception.

Audited every other bounds check in the file for the same overflow class
(section-table walk, PointerToRawData/SizeOfRawData, container count,
UX-container-size guard) - none of them perform addition/multiplication on
attacker-supplied values, so none are bypassable this way.

Rewrote the pinning test: it previously documented the bypass
(BypassesExplicitRangeCheck_ButStillFailsSafely, asserting the guard did NOT
fire); it now asserts the guard rejects directly
(IsRejectedByExplicitRangeCheck, asserting "out of range" IS present).

+++ src/FalkForge.Decompiler/WixBurnAccess.cs
@@ -57,7 +57,12 @@ internal sealed class WixBurnAccess : IWixBurnAccess
             stream.Seek(0x3C, SeekOrigin.Begin);
             var eLfanew = reader.ReadInt32();
 
-            if (eLfanew < 0 || eLfanew + 4 > stream.Length)
+            // Widen to long before adding: eLfanew is an attacker-controlled Int32 read
+            // straight from the file, and `eLfanew + 4` computed in unchecked Int32
+            // arithmetic can wrap to a negative value when eLfanew is near
+            // Int32.MaxValue, silently defeating this guard. stream.Length is already
+            // a long, so comparing in long arithmetic cannot overflow.
+            if (eLfanew < 0 || (long)eLfanew + 4 > stream.Length)
             {
                 reader.Dispose();
                 stream.Dispose();
